fixing trace function
The trace_level was being set in authenticated_with before the Api.api
variable was being reset so the value was being lost.  Moved the setter
and trace comment down below when Api.api is finally set so the trace
values are usable throughout the library.

All fixed small bug that was printing the string literal "text" instead
of the variable text in def trace().